Ecode_Camp_Task_Number_Five/Classify_Images_CNN_Django/classify_images_app/requirements.txt
Django
tensorflow
Pillow
numpy